Modify base to match oryx configuration
This commit gets us closer to the baseline of the Oryx that I started
with. I want to be able to replicate the lighting options I have with
the Oryx version, though, and I don't know the level of work that'll
entail.

The ultimate goal of this effort is to allow for the customization that
I can get from both the Oryx interface and the hand crafting through the
qmk firmware directly.