Set keymap to general baseline
So with this commit I'm getting the keymap to generally mimick the
keymap I am running with at the moment. The next steps will be to see
what it is like to work with the LEDs directly and to make sure I'm not
losing any of the cool lighting functionality I have in place right now.